HTCONDOR-2128 Add faster fd closing to my_popen and master restart as well
